refactored image reading routines such that alpha-channels are ommitted from double arrays and only inserted right before conversion to vtkImageData object (where this information is needed)

+++ Gui/DataView/GeoMapper.h
@@ -58,7 +58,7 @@ private:
 	double _cellsize;
 	unsigned _width;
 	unsigned _height;
-	float* _img_data;
+	double* _img_data;
 };
 
 #endif //GEOMAPPER_H
